fix(orchestration): release a proven exit only after its settlement is durable
The host keeps an observed exit until its owner has acted on it. Main
now says so: after the surface persists and the orchestration write
lands, it sends the same shutdown a closed pane sends, and the daemon
drops the record. A settlement that failed at either write sends
nothing, so the next sweep reads the same evidence. Reading is never
delivery; the owner's own act is.

The two daemon recovery cases move to part-08 to stay under the test
line cap.
