+++ src/main/runtime/orca-runtime.test.ts
@@ -57,6 +57,7 @@ await import('./orca-runtime-tests/terminal-output-and-worker-recovery-part-04.s
 await import('./orca-runtime-tests/terminal-output-and-worker-recovery-part-05.spec')
 await import('./orca-runtime-tests/terminal-output-and-worker-recovery-part-06.spec')
 await import('./orca-runtime-tests/terminal-output-and-worker-recovery-part-07.spec')
+await import('./orca-runtime-tests/terminal-output-and-worker-recovery-part-08.spec')
 await import('./orca-runtime-tests/terminal-handles-and-agent-status.spec')
 await import('./orca-runtime-tests/terminal-handles-and-agent-status-part-02.spec')
 await import('./orca-runtime-tests/terminal-settled-prompt-foreground-confirmation.spec')
